update rrt, file cleanup
