+++ mav_planning_benchmark/include/mav_planning_benchmark/global_planning_benchmark.h
@@ -6,7 +6,7 @@
 #include <mav_path_smoothing/velocity_ramp_smoother.h>
 #include <mav_planning_common/physical_constraints.h>
 #include <voxblox_ros/esdf_server.h>
-#include <voxblox_rrt_planner/voxblox_ompl_rrt.h>
+#include <voxblox_rrt_planner/ompl_rrt_voxblox.h>
 #include <voxblox_skeleton/io/skeleton_io.h>
 #include <voxblox_skeleton/skeleton_planner.h>
 #include <voxblox_skeleton/sparse_graph_planner.h>
