+++ voxblox_rrt_planner/include/voxblox_rrt_planner/ompl_rrt_voxblox.h
@@ -1,15 +1,16 @@
-#ifndef VOXBLOX_RRT_PLANNER_BLOX_OMPL_RRT_H_
-#define VOXBLOX_RRT_PLANNER_BLOX_OMPL_RRT_H_
+#ifndef VOXBLOX_RRT_PLANNER_VOXBLOX_OMPL_RRT_H_
+#define VOXBLOX_RRT_PLANNER_VOXBLOX_OMPL_RRT_H_
 
 #include <mav_msgs/conversions.h>
 #include <mav_msgs/eigen_mav_msgs.h>
 #include <ros/ros.h>
 
 #include "voxblox_rrt_planner/ompl/mav_setup.h"
+//#include <voxblox_rrt_planner/ompl_rrt_voxblox.h>
 
 namespace mav_planning {
 
-class BloxOmplRrt {
+class VoxbloxOmplRrt {
  public:
   enum RrtPlannerType {
     kRrtConnect = 0,
@@ -19,8 +20,8 @@ class BloxOmplRrt {
     kPrm
   };
 
-  BloxOmplRrt(const ros::NodeHandle& nh, const ros::NodeHandle& nh_private);
-  virtual ~BloxOmplRrt() {}
+  VoxbloxOmplRrt(const ros::NodeHandle& nh, const ros::NodeHandle& nh_private);
+  virtual ~VoxbloxOmplRrt() {}
 
   inline void setRobotRadius(double robot_radius) {
     robot_radius_ = robot_radius;
@@ -30,10 +31,8 @@ class BloxOmplRrt {
 
   // Both are expected to be OWNED BY ANOTHER OBJECT that shouldn't go out of
   // scope while this object exists.
-  /*
   void setTsdfLayer(voxblox::Layer<voxblox::TsdfVoxel>* tsdf_layer);
   void setEsdfLayer(voxblox::Layer<voxblox::EsdfVoxel>* esdf_layer);
-  */
 
   inline void setOptimistic(bool optimistic) { optimistic_ = optimistic; }
   bool getOptimistic() const { return optimistic_; }
@@ -47,7 +46,7 @@ class BloxOmplRrt {
   void setPlanner(RrtPlannerType planner) { planner_type_ = planner; }
 
   // Only call this once, only call this after setting all settings correctly.
-  virtual void setupProblem();
+  void setupProblem();
 
   // Fixed start and end locations, returns list of waypoints between.
   bool getPathBetweenWaypoints(
@@ -69,11 +68,6 @@ class BloxOmplRrt {
   void constructPrmRoadmap(double roadmap_construction_sec) {
     problem_setup_.setup();
     problem_setup_.constructPrmRoadmap(roadmap_construction_sec);
-  }
-
-  void getPrmRoadmap() {
-    ROS_INFO_STREAM("[BloxOmplRrt] getting roadmap");
-    problem_setup_.getPrmRoadmap();
   }
 
  protected:
@@ -108,34 +102,14 @@ class BloxOmplRrt {
   Eigen::Vector3d lower_bound_;
   Eigen::Vector3d upper_bound_;
 
+  // NON-OWNED pointers to the relevant layers. TSDF only used if optimistic,
+  // ESDF only used if pessimistic.
+  voxblox::Layer<voxblox::TsdfVoxel>* tsdf_layer_;
+  voxblox::Layer<voxblox::EsdfVoxel>* esdf_layer_;
+
   double voxel_size_;
-};
-
-class CbloxOmplRrt : public BloxOmplRrt {
- public:
-  CbloxOmplRrt(const ros::NodeHandle& nh, const ros::NodeHandle& nh_private)
-      : BloxOmplRrt(nh, nh_private) {};
-  virtual ~CbloxOmplRrt() {}
-
-  // Only call this once, only call this after setting all settings correctly.
-  void setupProblem(){
-    problem_setup_.setCbloxCollisionChecking(robot_radius_, voxel_size_,
-        map_distance_function_);
-    BloxOmplRrt::setupProblem();
-  }
-  void setMapDistanceCallback(
-      std::function<double(const Eigen::Vector3d& position)> function) {
-    map_distance_function_ = function;
-  }
-  void setVoxelSize(voxblox::FloatingPoint voxel_size) {
-    voxel_size_ = voxel_size;
-  }
-
- private:
-  // map of some sorts
-  std::function<double(const Eigen::Vector3d& position)> map_distance_function_;
 };
 
 }  // namespace mav_planning
 
-#endif  // VOXBLOX_RRT_PLANNER_BLOX_OMPL_RRT_H_
+#endif  // VOXBLOX_RRT_PLANNER_VOXBLOX_OMPL_RRT_H_
