+++ voxblox_rrt_planner/src/ompl_rrt_voxblox.cpp
@@ -1,4 +1,4 @@
-#include "voxblox_rrt_planner/voxblox_ompl_rrt.h"
+#include "voxblox_rrt_planner/ompl_rrt_voxblox.h"
 
 namespace mav_planning {
