+++ voxblox_rrt_planner/src/rrt_planner_voxblox.cpp
@@ -7,7 +7,7 @@
 #include <voxblox/utils/planning_utils.h>
 
 #include "voxblox_rrt_planner/rrt_planner.h"
-#include "voxblox_rrt_planner/voxblox_rrt_planner.h"
+#include "voxblox_rrt_planner/rrt_planner_voxblox.h"
 
 namespace mav_planning {
